move postgres client to apis
