+++ packages/brain/src/modules/apiClients/index.ts
@@ -3,5 +3,6 @@ export { BeaconchainApi as Beaconchain } from "./beaconchain.js";
 export { ValidatorApi } from "./validator.js";
 export { StandardApi } from "./standard.js";
 export { Web3SignerApi } from "./web3signer.js";
+export { PostgresClient } from "./postgres/index.js";
 export { DappnodeSignatureVerifier } from "./DappnodeSignatureVerifier.js";
 export { ApiError } from "./error.js";
